pyproject.toml
[tool.poetry]
name = "googlehacks"
version = "0.1.0"
description = ""
authors = ["Mykhailo Briazkalo <mbriazka@uwaterloo.ca>"]
readme = "README.md"
packages = [
    {include = "backend", from = "src"},
    {include = "config", from = "src"},
    {include = "database", from = "src"},
    {include = "gemini", from = "src"},
    {include = "google_apis", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.10"
black = "^24.4.2"
flake8 = "^7.1.0"
isort = "^5.13.2"
mypy = "^1.10.1"
google-api-python-client = "^2.137.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-jwt-extended = "^4.6.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
typed-ast = "^1.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
strict_optional = true
